Fix and improve advice management
- Previously, advice was added whenever the mode was enabled and removed
when it was disabled. However, advice is global so disabling advice in
one buffer messed with other buffers. Now, advice is added once when the
package is loaded and removed when the package is unloaded.

- Previously, `with-eval-after-load` was called over and over (each time
the mode was enabled in a new dired buffer), stacking up in
`after-load-alist`. Now, we simply add the advice up-front because
advice can be added to a symbol BEFORE the associated function has been
defined/loaded.

- The advice has been changed from "around" to "after" to simplify it.

- Renamed `nerd-icons--dired-functions-to-refresh` to
`nerd-icons-dired--functions-to-refresh` to comply with the package-name
prefix conventions.