Remove (most) dired advice and rely on hooks
The previous advice-based approach:

1. Required advising many functions.
2. Had some bugs around installing/removing advice (see #37).
3. Lead to some bugs because the advice was triggered before the buffer
was ready (#19).

This version mostly does away with the advice and uses the
`dired-after-readin-hook`, which is how most Dired extensions operate.
This hook is triggered:

1. After adding entries to the Dired buffer.
2. With the buffer narrowed to the new entries.

The icon overlays are given the `evaporate` property, so we don't even
need to bother with cleanup (except when disabling the mode).

The one function we still need to advise is `wdired-abort-changes`.
Unfortunately, this function restores the buffer text from a saved copy
and copying text drops the overlays.